p06_mllm_vision/requirements.txt
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.12.0
datasets>=2.20.0
huggingface_hub>=0.25.0
safetensors>=0.4.0
qwen-vl-utils>=0.0.2
Pillow>=10.0.0
torchvision>=0.16.0
numpy>=1.24.0
tqdm>=4.65.0
